requirements.txt
sentence-transformers
numpy
torch
pandas
datasets
rank_bm25
faiss-cpu